+++ package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "dotenv": "^8.0.0",
     "express": "^4.16.4",
+    "less": "^3.9.0",
     "node-fetch": "^2.6.0",
     "npm-run-all": "^4.1.5",
     "react": "^16.8.6",
