+++ src/Frontend/App.jsx
@@ -4,7 +4,7 @@ import './css/App.css';
 import NavBar from './component/NavBar/NavBar';
 import Twitter from './component/Twitter/Twitter';
 import RouteChecker from './component/RouteChecker/RouteChecker';
-import Cards from './component/Cards/Cards';
+import Gallery from './component/Gallery/Gallery.jsx';
 import UsefulLinks from './component/UsefulLinks/UsefulLinks';
 import AboutUs from './component/About/AboutUs';
 import Footer from './component/Footer/Footer_get';
@@ -20,7 +20,7 @@ class App extends Component {
       <div className="App">
         <NavBar />
         <RouteChecker />
-        <Cards />
+        <Gallery />
         <Twitter />
         <UsefulLinks />
         <AboutUs />
