+++ src/Frontend/component/Footer/Footer_get.jsx
@@ -1,6 +1,5 @@
 import React from 'react'; 
 import FooterCSS from './FooterCSS.css'; 
-
 import AccessLive from './Images/Disab logo.png';
 import GitHub from './Images/githublogo.svg';
 import Gmail from './Images/google-plus-logo.svg';
