+++ src/Frontend/component/Gallery/Cards/Cards.css
@@ -1,0 +1,82 @@
+
+.card {
+  position: relative;
+  background: white;
+  z-index: 1;
+
+  width: 336px;
+  height: 470px;
+  margin: 10px;
+  overflow:auto;
+
+  border-radius: .25rem;
+
+  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
+
+  transition: all 0.25s ease-in;
+  border-bottom: 5px solid transparent;
+}
+
+.card:hover{
+  transform: translateY(-5px);
+  border-bottom: 5px solid #2196fe;
+}
+
+.cardImg{
+  float:right;
+  height:auto;
+  position: relative;
+  size: cover;
+  background-position: center;
+}
+
+.cardTitle{
+    position: absolute;
+    bottom: 8px;
+    left: 75px;
+    color: tomato;
+    font-family: "Avenir Next", "Avenir", sans-serif;
+    padding: 10px 25px 10px 25px;
+  }
+
+.cardDetails{
+  padding: 10px 25px 10px 25px;
+  bottom: 0;
+  color: #808080;
+  font-family: 'proxima-nova-soft', sans-serif;
+
+}
+
+.readMore{
+  position: absolute;
+  bottom: 0px;
+  text-align: left;
+
+
+  
+  padding: 10px 25px 10px 25px;
+  border-radius: 0px 0px 4px 5px;
+  
+  border-top: 1px solid #e0e0e0;
+  background: #efefef;
+  width: 100%;
+}
+
+a {
+  text-decoration:none;
+  font-size: 17px;
+  text-align: center;
+  
+
+  font-weight: 500;
+  letter-spacing: 2px;
+
+  color: black;
+  text-transform: uppercase;
+  display: inline-block;
+
+}
+a:hover {
+  color:tomato;
+  transition: all .25s ease-out;
+}
