+++ src/Frontend/component/Gallery/Gallery.css
@@ -1,0 +1,3 @@
+.layout{
+    display: flex;
+}
