+++ src/Frontend/component/Gallery/Gallery.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import Cards from './Cards/Cards.jsx'
+import './Gallery.css';
+
+export default function Gallery() {
+  return (
+    <div className='layout'>
+      <Cards></Cards>
+      <Cards></Cards>
+
+    </div>
+  );
+}
